+++ eric-frontend/package.json
@@ -7,8 +7,7 @@
     "react": "^16.4.1",
     "react-calendar-material": "^0.1.8",
     "react-dom": "^16.4.1",
-    "react-scripts": "1.1.4",
-    "reactstrap": "^6.3.0"
+    "react-scripts": "1.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
