--- eric-frontend/src/booking/BookContainer.js
@@ -1,13 +1,0 @@
-import {BookCal} from './BookCal'
-import {BookInfo} from './BookInfo'
-
-
-export const BookContainer () {
-    render(){
-        return(
-            <BookCal/>
-            <BookInfo/>
-        );
-    }
-}
-
